Реализовывает Шаг 4 (получение факультета студента и студентов факультета)

+++ src/main/java/ru/hogwarts/school/controller/FacultyController.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.hogwarts.school.model.Faculty;
+import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.service.FacultyService;
 
 import java.util.Collection;
@@ -47,6 +48,10 @@ public class FacultyController {
 
     @DeleteMapping(path = "{id}")
     public ResponseEntity<Void> deleteFaculty(@PathVariable(value = "id") Long id) {
+        Faculty foundFaculty = facultyService.findFaculty(id);
+        if (foundFaculty == null) {
+            return ResponseEntity.notFound().build();
+        }
         facultyService.deleteFaculty(id);
         return ResponseEntity.ok().build();
     }
@@ -62,6 +67,11 @@ public class FacultyController {
             return ResponseEntity.ok(facultyService.findByNameOrColor(name, color));
         }
         return ResponseEntity.ok(facultyService.getAll());
+    }
+
+    @GetMapping(path = "{id}/students")
+    public ResponseEntity<Collection<Student>> getFacultyStudents(@PathVariable(name = "id") Long id) {
+        return ResponseEntity.ok(facultyService.getFacultyStudents(id));
     }
 
 }
